Move date validation to shared validates file
The validation is now dumb enough to be reusable
